defconfig: Enable BFQ io scheduler and mark it as default scheduler.
The BFQ I/O scheduler tries to distribute bandwidth among
all processes according to their weights.
It aims at distributing the bandwidth as desired, independently of
the disk parameters and with any workload. It also tries to
guarantee low latency to interactive and soft real-time
applications. If compiled built-in (saying Y here), BFQ can
be configured to support hierarchical scheduling.

Change-Id: I93e8068c3ab5f529df473db46d7701bb65eef2ec